cleanup of missing_number in Python and Perl
The location of the sum = 0 variable in the
Python and Perl version of missing_number has
been moved closer to its loop.  This more
closely matches the C version.

Output of the intermediate versions in the
Python missing_number program has been
removed.  The other examples do not have this.