separated python2 and python3
Previously the Python examples used 'python' which
defaulted to 'python2' under Debian.  This has now
been separated in to the more explicit python2
and python3.

Most examples remained unchanged but there were a
few differences.  raw_input change to input.
The binary manipulation done in the revbits example
works in python2 but requires significant changes to
work with python3.  This has not been completed yet.